Fixed Bugs and added right click Functionality
Things Added/Updated -

Made it so now there are no buttons on the right side of the task. You can right click and perform the taks

Bugs Fixed -

Typing a long task would go out of the container box. now it will wrap and start from next line
Fixed when clicking on edit and reloading would save the task as empty.