The remaking of the entire website
I took this decision because the older version i took so much help of chat gpt but this time i will do it myself. i will get guided if i need i will google but to guide myself not to just get a ready made solution and copy paste it here.