Fixed Vercel Deployment
Took me hours and hours to fix this. And i am sure if i was a skilled developer i would have fixed this in minutes cause there was just a few changes of codes which made it work. like literally. Anyways. I will remember this and this probably wont happen in the future.